+++ storage-proofs/src/drgraph.rs
@@ -84,7 +84,7 @@ pub trait Graph<H: Hasher>: ::std::fmt::Debug + Clone + PartialEq + Eq {
     ///
     /// The `parents` parameter is used to store the result. This is done fore performance
     /// reasons, so that the vector can be allocated outside this call.
-    fn parents(&self, node: usize, parents: &mut [usize]);
+    fn parents(&self, node: usize, parents: &mut [u32]);
 
     /// Returns the size of the graph (number of nodes).
     fn size(&self) -> usize;
@@ -101,7 +101,7 @@ pub trait Graph<H: Hasher>: ::std::fmt::Debug + Clone + PartialEq + Eq {
         &self,
         id: &H::Domain,
         node: usize,
-        parents: &[usize],
+        parents: &[u32],
         parents_data: &[u8],
         exp_parents_data: Option<&[u8]>,
     ) -> Result<Self::Key>;
@@ -143,7 +143,7 @@ impl<H: Hasher> Graph<H> for BucketGraph<H> {
         &self,
         id: &H::Domain,
         node: usize,
-        parents: &[usize],
+        parents: &[u32],
         base_parents_data: &[u8],
         _exp_parents_data: Option<&[u8]>,
     ) -> Result<Self::Key> {
@@ -151,9 +151,9 @@ impl<H: Hasher> Graph<H> for BucketGraph<H> {
         hasher.update(AsRef::<[u8]>::as_ref(id));
 
         // The hash is about the parents, hence skip if a node doesn't have any parents
-        if node != parents[0] {
+        if node != parents[0] as usize {
             for parent in parents.iter() {
-                let offset = data_at_node_offset(*parent);
+                let offset = data_at_node_offset(*parent as usize);
                 hasher.update(&base_parents_data[offset..offset + NODE_SIZE]);
             }
         }
@@ -163,7 +163,7 @@ impl<H: Hasher> Graph<H> for BucketGraph<H> {
     }
 
     #[inline]
-    fn parents(&self, node: usize, parents: &mut [usize]) {
+    fn parents(&self, node: usize, parents: &mut [u32]) {
         let m = self.degree();
 
         match node {
@@ -198,15 +198,15 @@ impl<H: Hasher> Graph<H> for BucketGraph<H> {
 
                     // remove self references and replace with reference to previous node
                     if out == node {
-                        *parent = node - 1;
+                        *parent = (node - 1) as u32;
                     } else {
                         assert!(out <= node);
-                        *parent = out;
+                        *parent = out as u32;
                     }
                 }
 
                 // Add the immediate predecessor as a parent to ensure unique topological ordering.
-                parents[m_prime] = node - 1;
+                parents[m_prime] = (node - 1) as u32;
             }
         }
     }
@@ -297,7 +297,7 @@ mod tests {
 
                 for parent in p1 {
                     // TODO: fix me
-                    assert_ne!(i, parent, "self reference found");
+                    assert_ne!(i, parent as usize, "self reference found");
                 }
             }
         }
